% Auto-generated: do not edit by hand
\name{dashDashSign}

\alias{dashDashSign}

\title{DashSign component}

\description{
DashSign is a canvas for capturing hand written signatures
}

\usage{
dashDashSign(id=NULL, backgroundColor=NULL, data=NULL, n_submit=NULL,
penColor=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{backgroundColor}{Character. The background color of the signature canvas}

\item{data}{Character. Base 64 data string for the resulting image}

\item{n_submit}{Numeric. The number of times the save button has been clicked}

\item{penColor}{Character. The pen color of the signature component}

\item{style}{Named list. The style props of signature div}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

